# application/services/custom_services/products/src/models/entitymodels/products.ts
export interface products 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   modules: { type: String, ref: 'modules' },
   ienum: { type: Ienum }
}


export enum Ienum {
   OPEN = 'open',
   CLOSE = 'close',
   INPROGRESS = 'inprogress'
}